+++ packages/azure-openapi-validator/autorest/src/spectral-plugin-func.ts
@@ -80,10 +80,7 @@ export async function spectralPluginFunc(initiator: IAutoRestPluginInitiator): P
       const normalizedFile = file.startsWith("file:///") ? fileURLToPath(file) : file
       await runSpectral(openapiDefinitionObject, normalizedFile, initiator.Message.bind(initiator), spectral)
     } catch (e) {
-      initiator.Message({
-        Channel: "fatal",
-        Text: `spectralPluginFunc: Failed validating: '${file}', error encountered: ` + e,
-      })
+      catchSpectralRunErrors(file, e, initiator)
     }
   }
 
@@ -119,6 +116,18 @@ function ifNotStagingRunThenDisableStagingOnlyRules(
         // Resulting in value of -1, as defined here:
         // https://github.com/stoplightio/spectral/blob/44c40e2b7c9ea6222054da8700049b0307cc5f8b/packages/ruleset-migrator/src/transformers/rules.ts#L39
         rule.severity = "off"
+        // We must disable the rule to ensure it is not run at all. Spectral source for that is in [1].
+        // Otherwise, if it throws due to a bug in rule implementation, it will result in fatal error.
+        // Example of how a rule can thrown is given in [2].
+        // Example where we disable a rule from running in production so it doesn't throw is given in [3].
+        // Without this line, the fix in [3] doesn't help. For a proof of that, see [4] and its log, [5],
+        // showing that a rule that is supposed to run in stagingOnly, still throws in production.
+        // [1] https://github.com/stoplightio/spectral/blob/6d0991572f185ce5cf4031dc1d8eb4035b5eaf1d/packages/core/src/runner/runner.ts#L39
+        // [2] https://github.com/Azure/azure-openapi-validator/pull/595
+        // [3] https://github.com/Azure/azure-openapi-validator/pull/596
+        // [4] https://github.com/Azure/azure-rest-api-specs/pull/26024
+        // [5] https://dev.azure.com/azure-sdk/internal/_build/results?buildId=3125612&view=logs&j=0574a2a6-2d0a-5ec6-40e4-4c6e2f70bea2&t=80c3e782-49f0-5d1c-70dd-cbee57bdd0c7&l=252
+        rule.enabled = false
       }
     })
   }
@@ -232,4 +241,30 @@ export async function getRuleSet(openapiType: OpenApiTypes): Promise<[ruleset: R
     rulesInStagingOnlyByName.forEach(([_, rule]) => delete rule.stagingOnly)
     return namesOfRulesInStagingOnly
   }
+}
+
+/** Spectral (from "@stoplight/spectral-core") may throw https://www.npmjs.com/package/es-aggregate-error
+ * If so, we print out all the constituent errors.
+ * For additional context, see: https://github.com/Azure/azure-sdk-tools/issues/6856
+ */
+function catchSpectralRunErrors(file: string, e: any, initiator: any) {
+  // Initialize an array to collect error messages
+  const errorMessages: string[] = [e]
+
+  // Check if "e" contains the "errors" property
+  if (e && e.errors && Array.isArray(e.errors)) {
+    e.errors.forEach((error: any, index: number) => {
+      // Push each error message into the array
+      errorMessages.push(`Error ${index + 1}: ${error.message}`)
+    })
+  }
+
+  // Combine all error messages with newlines
+  const combinedErrorMessages = errorMessages.join("\n")
+
+  // Call initiator.Message with the combined error message
+  initiator.Message({
+    Channel: "fatal",
+    Text: `spectralPluginFunc: Failed validating: '${file}'. Errors encountered:\n${combinedErrorMessages}`,
+  })
 }