+++ packages/rulesets/src/spectral/functions/tracked-extension-resources-are-not-allowed.ts
@@ -1,0 +1,36 @@
+import { isPathOfExtensionResource, getProperty } from "./utils"
+
+export const trackedExtensionResourcesAreNotAllowed = (apiPath: any, _opts: any, ctx: any) => {
+  if (apiPath === null || typeof apiPath !== "string") {
+    return []
+  }
+
+  const verbs = ["get", "put", "patch"]
+  const responseCodes = ["200", "201", "202"]
+  const path = ctx.path || []
+  const errors = []
+
+  // check if the apiPath is of type extensionResource
+  if (isPathOfExtensionResource(apiPath)) {
+    const operationPaths = ctx?.documentInventory?.resolved?.paths[apiPath]
+    for (const verb of verbs) {
+      if (operationPaths[verb]) {
+        //check if location property exists in the response schema
+        for (const responseCode of responseCodes) {
+          const responseSchema = operationPaths[verb].responses[responseCode]?.schema
+          if (responseSchema) {
+            const locationProperty = getProperty(responseSchema, "location")
+            if (locationProperty) {
+              errors.push({
+                message: `${apiPath} is an extension resource and ${responseCode} response schema in ${verb} operation includes location property. Extension resources of type tracked are not allowed.`,
+                path: [...path, verb, responseCode],
+              })
+            }
+          }
+        }
+      }
+    }
+  }
+
+  return errors
+}